requirements.txt
pydantic==2.7.1
pandas==2.2.2
torch==2.3.1
torchvision==0.18.1
llama-cpp-python==0.2.80
instructor==1.3.3
transformers==4.41.2
scikit-learn==1.5.1
sentencepiece